+++ velox/experimental/stateful/state/RocksDBKeyedStateBackend.cpp
@@ -178,10 +178,10 @@ RocksDBKeyedStateBackend::getOrCreateRankMapState(
   return nullptr;
 }
 
-std::shared_ptr<InternalTimerService<uint32_t, int64_t>>
+std::shared_ptr<InternalTimerService<int64_t, int64_t>>
 RocksDBKeyedStateBackend::createTimerService(
-    Triggerable<uint32_t, int64_t>* triggerable) {
-  return std::make_shared<InternalTimerService<uint32_t, int64_t>>(triggerable);
+    Triggerable<int64_t, int64_t>* triggerable) {
+  return std::make_shared<InternalTimerService<int64_t, int64_t>>(triggerable);
 }
 
 std::shared_ptr<InternalTimerService<uint32_t, TimeWindow>>
